Configured server not to serve the special backup file, separated CRC32 functionality and finalized the fraction sorting.
